requirements.txt
# 图像处理相关
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# 深度学习框架
torch>=2.0.0

# 模型和AI相关
modelscope>=1.9.0

# 网络请求
requests>=2.31.0

# 视频处理
imageio-ffmpeg>=0.4.8

# ComfyUI相关（通常由ComfyUI环境提供，这里仅作记录）
# comfy
# folder_paths
# nodes
